+++ CHANGELOG.md
@@ -5,13 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.10.3] - 1024-12-27
+## [0.11.2] - 2024-12-30
+
+- Fix SyntaxError: f-string expression part cannot include a backslash.
+
+## [0.11.1] - 2024-12-29
+
+- Improve prompts.
+- Fix issue #24.
+
+## [0.11.0] - 2024-12-28
+
+- Add CiaynAgent to support models that do not have, or are not good at, agentic function calling.
+- Improve env var validation.
+- Add --temperature CLI parameter.
+
+## [0.10.3] - 2024-12-27
 
 - Fix logging on interrupt.
 - Fix web research prompt.
 - Simplify planning stage by executing tasks directly.
 - Make research notes available to more agents/tools.
-- Make read_file always output status panel.
 
 ## [0.10.2] - 2024-12-26